Call check_disk_change(), and have dex_media_changed() always be true
The .media_changed function is only used by check_disk_change(), which
is meant to be called by *us* the first time a device is opened.  So we
do that.

Since we can't reliably tell whether a card was swapped while we weren't
looking, we go ahead and pretend this is always the case.  I'm not
exactly sure whether we need to do this (heck, I'm not even sure it
*does* anything), but it cannot hurt, as it's only called when opening a
device that was previously closed.