Merging (and shrinking) both input and output buffers into one
There is no need for separate input and output buffers, since
communication with the DexDrive is half-duplex.  This merges them
together, taking care not to overwrite pending outbound data with
inbound.

While we're at it, this also shrinks the buffer from 1024 to 261 bytes,
to avoid wasting memory.  (Not that anyone will notice, but still.)