Replacing the thread and bio list with a simple work queue
Instead of managing our own thread and list to handle bios, we simply
create a work queue and let the kernel deal with the details.  Not only
is this simpler (aside from the need to create a container), it will
also allow us to perform initialization much sooner eventually.  It
should also make it easier, later on, to ensure that this driver can
suspend properly.